third_party/blink/web_tests/external/wpt/web-animations/animation-model/animation-types/addition-per-property-002-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] stroke-dasharray: dasharray
  assert_equals: The value should be 1px, 2px, 3px, 4px, 5px at 0ms expected "1px, 2px, 3px, 4px, 5px" but got "7px, calc(30% + 2px), 5px, 10px, calc(30% + 5px), 3px, 8px, calc(30% + 3px), 6px, 11px, calc(30% + 1px), 4px, 9px, calc(30% + 4px), 7px"
[FAIL] text-decoration: "overline" onto "underline"
  assert_equals: The value should be overline at 0ms expected "overline" but got "overline solid rgb(0, 0, 0)"
[FAIL] text-decoration: "underline" onto "overline"
  assert_equals: The value should be underline at 0ms expected "underline" but got "underline solid rgb(0, 0, 0)"
[FAIL] word-spacing: percentage
  assert_equals: The value should be 130% at 0ms expected "130%" but got "0px"
[FAIL] word-spacing: units "%" onto "px"
  assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "10px"
[FAIL] word-spacing: units "px" onto "%"
  assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "10px"
[FAIL] word-spacing: units "rem" onto "%"
  assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "20px"
[FAIL] word-spacing: units "%" onto "rem"
  assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "20px"
[FAIL] word-spacing: units "calc" onto "px"
  assert_equals: The value should be calc(20% + 30px) at 0ms expected "calc(20% + 30px)" but got "10px"
[FAIL] word-spacing: calc
  assert_equals: The value should be calc(30% + 30px) at 0ms expected "calc(30% + 30px)" but got "0px"
Harness: the test ran to completion.
